Use separate test classes for testing.
