+++ src/live-region/test-classes/styles.scss
@@ -1,0 +1,12 @@
+/*
+ Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ SPDX-License-Identifier: Apache-2.0
+*/
+
+.root {
+  /* used in test-utils or tests */
+}
+
+.announcer {
+  /* used in test-utils or tests */
+}
